requirements.txt
# ── Web Framework ────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# ── Validation / Settings ────────────────────────────────────────
pydantic==2.8.0
pydantic-settings==2.4.0

# ── Agent Framework (LangGraph — NO LangChain Agent) ─────────────
langgraph==0.2.28
langchain-core==0.3.10        # chỉ dùng Message types & BaseTool interface

# ── LLM Providers (multi-provider, swap dễ dàng) ─────────────────
anthropic==0.34.0
openai==1.45.0                # GPT-4o + Gemini via OpenAI-compat + embedding
google-generativeai==0.8.1    # Gemini native
# ollama → gọi qua openai-compat (localhost:11434/v1) → không cần package

# ── Vector DB ────────────────────────────────────────────────────
qdrant-client==1.11.0

# ── Document Parsing ─────────────────────────────────────────────
pypdf==4.3.0
python-docx==1.1.2
openpyxl==3.1.5
Pillow==10.4.0
pytesseract==0.3.13
unstructured[pdf,docx]==0.15.0

# ── Storage / DB ─────────────────────────────────────────────────
sqlalchemy==2.0.35
alembic==1.13.2
asyncpg==0.29.0
aiofiles==24.1.0

# ── Cache / Rate-limit ───────────────────────────────────────────
redis==5.0.8

# ── HTTP Client (SQL API backend) ────────────────────────────────
httpx==0.27.0
tenacity>=8.2.0,<9.0.0

# ── Security ─────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.0

# ── Logging ──────────────────────────────────────────────────────
structlog==24.4.0

# ── Tokenizer ────────────────────────────────────────────────────
tiktoken==0.7.0

# ── Dev / Test ───────────────────────────────────────────────────
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-mock==3.14.0
httpx==0.27.0